backend/requirements.txt
Flask
Flask-Cors
Flask-SQLAlchemy
Flask-Migrate
marshmallow
python-dotenv
PyMySQL
